converted allEvents function to EventsController class, cleaned consols, deleted unneeded files, and make some changes relates #13
