--- src/controllers/Events/index.ts
@@ -1,3 +1,0 @@
-import allEvents from './allEvents'
-
-export { allEvents }
