--- src/controllers/index.ts
@@ -1,3 +1,0 @@
-import { allEvents } from "./Events";
-
-export {allEvents}
