+++ src/db/build.ts
@@ -1,7 +1,6 @@
 import { sequelize, Event, User } from '.'
 import fakeData from './FakeData/fakeData.json'
-
-const { NODE_ENV } = process.env
+import config from '../config/environment'
 
 const build = async () => {
   await sequelize.sync({ force: true })
@@ -9,9 +8,8 @@ const build = async () => {
   await User.bulkCreate(fakeData.Users)
 }
 
-if (NODE_ENV !== 'test') {
-  console.log('success')
-  build()
+if (config.nodeEnv !== 'test') {
+  build().then(() => console.log('database built successfully'))
 }
 
 export default build
