+++ test/filterEvents.test.ts
@@ -7,7 +7,7 @@ import sequelize from '../src/db/connection'
 beforeEach(() => build())
 
 describe('Filter events tests', () => {
-  test('length of all products', (done) => {
+  test('length of all events', (done) => {
     supertest(app)
       .get('/api/v1/events')
       .expect('Content-Type', /json/)
@@ -18,29 +18,25 @@ describe('Filter events tests', () => {
       })
   })
 
-  test('get all events filtered by status:closed', (done) => {
+  test('test an event filtered by status:closed', (done) => {
     supertest(app)
       .get('/api/v1/events/?status=closed')
       .expect('Content-Type', /json/)
       .end((err, res) => {
         if (err) return done(err)
-        console.log(res.body)
-
-        expect(res.body[0].description).toEqual(
-          'you should wear a custom'
+        expect(res.body[0].status).toEqual(
+          'closed'
         )
         return done()
       })
   })
 
-  test('get all products filtered by data', (done) => {
+  test('get an event filtered by date', (done) => {
     supertest(app)
       .get('/api/v1/events/?from=2022-1-20&to=2022-1-28')
       .expect('Content-Type', /json/)
       .end((err, res) => {
         if (err) return done(err)
-        console.log(res.body)
-
         expect(res.body[0].name).toEqual(
           'Hiking'
         )
@@ -49,4 +45,4 @@ describe('Filter events tests', () => {
   })
 })
 
-afterAll(() => sequelize.close)
+afterAll(() => sequelize.close())
