--- test/main.test.ts
@@ -1,8 +1,0 @@
-// import { describe, expect, test } from '@jest/globals'
-
-describe('Server.ts tests', () => {
-  test('Math test', () => {
-    expect(2 + 2).toBe(4)
-  })
-})
-export {}
